requirements.txt
langchain
langchain-community
faiss-cpu
sentence-transformers
streamlit
google-generativeai